Add simple ldap authentication
auth.php allows to submit an Email address and a password

inc/authenticate.php looks up the user with that Email address and tries to
bind with the given password. The Email address - if found - is also
checked against a list of authorized users.

Authenticated and authorized users get a session.

All web facing php files include inc/session.php and redirect to auth.php
if the user is not in the session.

Credits go to: https://samjlevy.com/php-ldap-login/